fixing the problem that import command wasnt working because Ecto wasnt being initialized in the mix command

+++ priv/repo/migrations/20180107155720_create_stock_prices.exs
@@ -16,7 +16,7 @@ defmodule Investtrack.Repo.Migrations.CreateStockPrices do
       add :best_buy, :decimal
       add :best_sell, :decimal
       add :trade_count, :integer
-      add :trade_amount, :integer
+      add :trade_amount, :bigint
       add :trade_volume, :decimal
       add :date, :naive_datetime, primary_key: true
